Fix Lokasenna_Create folder to contain selected tracks.lua
[The current implementation includes the following line](https://github.com/ReaTeam/ReaScripts/blob/7c05f6416cc0625117250cda6264d77053daf321/Tracks%20Properties/Lokasenna_Create%20folder%20to%20contain%20selected%20tracks.lua#L44):

```lua
local newDepth = (curDepth == 0) and (curDepth - 1) or -1
```

This line doesn’t make sense, because if `curDepth == 0` holds, then `curDepth - 1` is `-1`, so `newDepth` is always `-1`.

I believe the intent is to say `newDepth = curDepth - 1`, which means the that the `lastSel` is now the last track of yet another folder (the one we just created).

With this fix, the action behaves more sanely when creating folders within folders. Consider the following example:

```
TRACK 1
  TRACK 2 (SELECTED)
  TRACK 3 (SELECTED)
TRACK 4
```

In the current implementation the result would be:

```
TRACK 1
  TRACK NEW
    TRACK 2 (SELECTED)
    TRACK 3 (SELECTED)
  TRACK 4
```

But with my fix the result is:

```
TRACK 1
  TRACK NEW
    TRACK 2 (SELECTED)
    TRACK 3 (SELECTED)
TRACK 4
```

Note how `TRACK 4` hasn’t been nested into `TRACK 1`.

There’s still (both in the original version as well as in my fixed version) an inconvenience when the last selected track is itself a folder, for example:

```
TRACK 1 (SELECTED)
TRACK 2 (SELECTED)
  TRACK 3
```

Which results in:

```
TRACK NEW
  TRACK 1 (SELECTED)
  TRACK 2 (SELECTED)
TRACK 3
```

Note how `TRACK 2` is no longer a parent of `TRACK 3`.

But fixing this would require traversing `TRACK 3` onward looking for the last child. This would complicate the action beyond its scope (I think, if you disagree let me know and I can try to come up with a pull request). One can always select `TRACK 3` as well and then it works.